Moved validateApiList to utils directory
also made veteranFacing required in productionRequest.ts
added email() validation to primaryContact.email and
secondaryContact.email fields.
