+++ routes/DeveloperApplication.ts
@@ -8,24 +8,8 @@ import OktaService from '../services/OktaService';
 import GovDeliveryService from '../services/GovDeliveryService';
 import SlackService from '../services/SlackService';
 import DynamoService from '../services/DynamoService';
-import { API_LIST } from '../config/apis';
+import { validateApiList } from '../util/validators';
 import { DevPortalError } from '../models/DevPortalError';
-
-export function validateApiList(val: string): string {
-  let result: boolean;
-  try {
-    const apis = val.split(',');
-    result = apis.every(api => API_LIST.includes(api));
-  } catch {
-    throw new Error('it was unable to process the provided data');
-  }
-
-  if (!result) {
-    throw new Error('invalid apis in list');
-  }
-
-  return val;
-}
 
 export const applySchema = Joi.object().keys({
   firstName: Joi.string().required(),
